Updated Index!
I wrote the first page from the website. It's a small description about who I am and what this is and then there is already my 4 Projects I want to show. I still need to do those 4 pages for the projects, and probably put a picture to the projects to make it a bit visual pleasing.